server/requirements.txt
fastapi==0.68.0
uvicorn==0.14.0
tortoise-orm==0.17.6
asyncpg==0.24.0
